requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
asyncpg
pydantic
python-jose
passlib
bcrypt
python-multipart
Pillow
fastapi-mail
alembic
captcha
requests
loguru
sse_starlette
apscheduler
etl-parser
cryptography
python-dotenv